Beginning minor changes to make major changes, since we're making a password manager now

+++ src/generator.cc
@@ -1,6 +1,6 @@
 #include "generator.hh"
 
-std::string generatePassword(int length, const std::string& characterSet) {
+std::string generatePassword(size_t length, const std::string& characterSet) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::uniform_int_distribution<> dis(0, characterSet.size() - 1);
